+++ cmd/infra_get_kubeconfig.go
@@ -12,6 +12,9 @@ var infraKubeconfigCmd = &cobra.Command{
 	Use:   "get-kubeconfig",
 	Short: "Output kubeconfig path",
 	Long:  "Output kubeconfig path",
+	Example: `To output kubeconfig path, run:
+
+    $ safira infra get-kubeconfig`,
 	RunE:  runGetKubeconfig,
 	SuggestionsMinimumDistance: 1,
 }