+++ cmd/template_pull.go
@@ -11,6 +11,9 @@ var templatePullCmd = &cobra.Command{
 	Use:   "pull",
 	Short: "Pull out the official Vertigo iPaaS templates",
 	Long:  "Pull out the official Vertigo iPaaS templates",
+	Example: `To pull the official templates, run:
+
+    $ safira template pull`,
 	RunE:  runTemplatePull,
 }